JWA-46 Captain actions validation added

+++ src/main/java/ru/rsreu/harbor/controller/command/captain/cancel_arrival/CancelArrivalRequestCommandLogicDbImpl.java
@@ -1,21 +1,34 @@
 package ru.rsreu.harbor.controller.command.captain.cancel_arrival;
 
+import ru.rsreu.harbor.controller.validation.CaptainActionValidator;
 import ru.rsreu.harbor.datalayer.dao.PierAssignmentDao;
 import ru.rsreu.harbor.datalayer.dao.UserDao;
+import ru.rsreu.harbor.datalayer.model.User;
 
 public class CancelArrivalRequestCommandLogicDbImpl implements CancelArrivalRequestCommandLogic {
     private final UserDao userDao;
     private final PierAssignmentDao pierAssignmentDao;
 
-    public CancelArrivalRequestCommandLogicDbImpl(UserDao userDao, PierAssignmentDao pierAssignmentDao) {
+    private final CaptainActionValidator captainActionValidator;
+
+    public CancelArrivalRequestCommandLogicDbImpl(
+            UserDao userDao,
+            PierAssignmentDao pierAssignmentDao,
+            CaptainActionValidator captainActionValidator) {
         this.userDao = userDao;
         this.pierAssignmentDao = pierAssignmentDao;
+        this.captainActionValidator = captainActionValidator;
     }
 
     @Override
     public void deleteRequestByCaptain(String captainLogin) {
-        this.pierAssignmentDao.delete(this.pierAssignmentDao.findByCaptain(
-                        this.userDao.findByLogin(captainLogin).orElseThrow(IllegalArgumentException::new))
-                .orElseThrow(IllegalArgumentException::new));
+        User captain = this.userDao.findByLogin(captainLogin).orElseThrow(IllegalArgumentException::new);
+        if (this.captainActionValidator.isCaptainActionValid(captain, null)) {
+            this.pierAssignmentDao.delete(this.pierAssignmentDao.findByCaptain(
+                            this.userDao.findByLogin(captainLogin).orElseThrow(IllegalArgumentException::new))
+                    .orElseThrow(IllegalArgumentException::new));
+        } else {
+            throw new IllegalArgumentException();
+        }
     }
 }
